pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vector-engine"
version = "0.1.0"
description = "ML-first vector computation and retrieval engine."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Vector Engine Contributors" }
]
dependencies = [
  "numpy>=1.24",
]

[project.optional-dependencies]
faiss = [
  "faiss-cpu>=1.7.4; platform_system != 'Darwin' or platform_machine != 'arm64'",
]
ml = [
  "scikit-learn>=1.3",
]
dev = [
  "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["vector_engine*"]
